[QOL] Fix DebugMode & Packet Order
* ObjAiBase:
  * Moved SetPosition to AttackableUnit.
* AttackableUnit:
  * Added SetPosition with optional repathing.
  * Finishing a path of waypoints no longer calls a StopMovement, instead simply clears waypoints.
* Chat Commands:
  * Fixed DebugMode attaching particles to missiles instead of on the ground.
* Packets:
  * HandleSpawn sends EnterVisibility for champions before buildings.
  * All S2C_ packets are ordered properly.
  * Added ToggleInputLockFlag packet to interface.
  * FXCreateGroup no longer causes a crash with null Caster.
* Scripting:
  * SummonerFlash particles spawn before teleporting.

+++ PacketDefinitions420/PacketExtensions.cs
@@ -173,7 +173,7 @@ namespace PacketDefinitions420
 
                         md = new MovementDataNormal
                         {
-                            SyncID = (int)unit.SyncId,
+                            SyncID = unit.SyncId,
                             TeleportNetID = unit.NetId,
                             // TODO: Implement teleportID (likely to be the index (starting at 1) of a waypoint we want to TP to).
                             // Crucial in syncing client positions with server positions, especially when entering vision
